+++ src/main/java/com/example/ecopr/repository/CategoriesRepository.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.ecopr.model.Categories;
 import java.util.Optional;
 
-public interface CategoriesRepository extends JpaRepository<Categories, Long> {
+public interface CategoriesRepository extends JpaRepository<Categories, Integer> {
     Optional<Categories> findByCategory(String category);
 }
