+++ src/main/java/com/example/ecopr/repository/ProductsRepository.java
@@ -2,6 +2,6 @@ package com.example.ecopr.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.ecopr.model.Products;
 
-public interface ProductsRepository extends JpaRepository<Products,Long>{
+public interface ProductsRepository extends JpaRepository<Products,Integer>{
     
 }
