+++ src/main/java/com/example/ecopr/repository/RelationsRepository.java
@@ -5,5 +5,5 @@ import com.example.ecopr.model.Relations;
 import java.util.List;
 
 public interface RelationsRepository extends JpaRepository<Relations, Long> {
-    List<Relations> findByCategoryId(Long categoryId);
+    List<Relations> findByCategoryId(Integer categoryId);
 }
